Fix PTY rendering with vt10x terminal emulator
- Add vt10x for proper VT100 terminal emulation
- Fix key-to-bytes conversion for proper PTY input
- Add ready state to wait for window size
- Improve resize handling

+++ internal/pty/manager.go
@@ -84,7 +84,7 @@ func (m *Manager) Write(id string, data []byte) (int, error) {
 }
 
 // ResizeAll resizes all sessions
-func (m *Manager) ResizeAll(rows, cols uint16) {
+func (m *Manager) ResizeAll(rows, cols int) {
 	m.mu.RLock()
 	defer m.mu.RUnlock()
